requirements.txt
gradio==3.39.1
google-generativeai==0.5.2
python-dotenv==1.0.1
